requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pymongo==4.13.0
pytesseract==0.3.13
python-dotenv==1.1.0
Werkzeug==3.1.3
